+++ src/frontend/assets/staticdata/static.mytbtranslations-de.js
@@ -535,6 +535,7 @@ window.importStaticTranslationsJsonP = `
     "label.odimgobject.detector.tfjs_mobilenet_v1": "mobilenet_v1",
     "label.odimgobject.detector.tfjs_posenet": "posenet",
     "label.odimgobject.detector.faceapi": "faceapi",
+    "label.odimgobject.detector.manual": "manuell",
     "label.odimgobject.detector.picasafile": "picasafile",
     "label.odimgobject.detector.picasa": "picasa",
     "label.odimgobject.detector.undefined": "Unbekannt",